refactor: update test
